debug-scripts/V4-FoundryProjectV1/requirements.txt
azure-ai-evaluation==1.11.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0
azure-ai-agents==1.2.0b5
azure-cosmos==4.9.0
azure-identity==1.24.0
azure-monitor-events-extension==0.1.0
azure-monitor-opentelemetry==1.7.0
azure-search-documents==11.5.3
fastapi==0.116.1
openai==1.105.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-openai==0.46.2
opentelemetry-sdk==1.36.0
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-cov==5.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
semantic-kernel==1.35.3
uvicorn==0.35.0
pylint-pydantic==0.3.5
pexpect==4.9.0
mcp==1.13.1
agent-framework>=1.0.0b251105
pyodbc
ipykernel
pydantic
